requirements.txt
langchain
streamlit
langchain_groq
python-dotenv
langchain_community
langchain_core
torch
transformers
librosa
streamlit_mic_recorder
chromadb
fpdf
pyttsx3
pypdfium2
flask
flask-cors
gunicorn 
flask_sqlalchemy
flask_login
flask_wtf
wtforms
diffusers
langchain_experimental
pymongo
PyPDF2
langchain_mongodb
pymupdf
pytesseract
pdf2image
pillow
altair==4
tiktoken==0.4.0